fix: integer promotion to avoid overflow

+++ src/shapes.rs
@@ -19,8 +19,8 @@ pub struct Polygon {
 
 impl Polygon {
     pub fn read_vertices(stream: &mut Cursor<Vec<u8>>, zoom: u16) -> Result<Polygon, io::Error> {
-        let bbw = stream.read_u8()? as i16 * zoom as i16 / 64;
-        let bbh = stream.read_u8()? as i16 * zoom as i16 / 64;
+        let bbw = (stream.read_u8()? as i32 * zoom as i32 / 64) as i16;
+        let bbh = (stream.read_u8()? as i32 * zoom as i32 / 64) as i16;
         let num_points = stream.read_u8()? as usize;
 
         assert!(num_points % 2 == 0, "Points must be even");
@@ -28,8 +28,8 @@ impl Polygon {
 
         let mut points: Vec<Point> = Vec::with_capacity(num_points);
         for _ in 0..num_points {
-            let x = stream.read_u8()? as i16 * zoom as i16 / 64;
-            let y = stream.read_u8()? as i16 * zoom as i16 / 64;
+            let x = (stream.read_u8()? as i32 * zoom as i32 / 64) as i16;
+            let y = (stream.read_u8()? as i32 * zoom as i32 / 64) as i16;
             points.push(Point { x, y });
         }
         Ok(Polygon { bbw, bbh, points })
